+++ tawazi/node/__init__.py
@@ -1,6 +1,8 @@
 """Node related Classes and functions."""
+from . import extend  # noqa: F401
 from .functions import ReturnUXNsType, wrap_in_uxns
-from .node import Alias, ArgExecNode, ExecNode, LazyExecNode, UsageExecNode
+from .node import Alias, ArgExecNode, ExecNode, LazyExecNode
+from .uxn import UsageExecNode
 
 __all__ = [
     "Alias",
