+++ tawazi/node/node.py
@@ -1,16 +1,16 @@
 """Module describing ExecNode Class and subclasses (The basic building Block of a DAG."""
 import functools
 import warnings
-from copy import copy, deepcopy
+from copy import copy
 from dataclasses import dataclass, field
-from functools import partial, reduce
+from functools import partial
 from threading import Lock
 from types import MethodType
-from typing import Any, Callable, Dict, Generic, List, NoReturn, Optional, Tuple, Union
+from typing import Any, Callable, Dict, Generic, List, Optional, Tuple, Union
 
 from loguru import logger
 
-from tawazi._helpers import _filter_noval, _make_raise_arg_error
+from tawazi._helpers import _make_raise_arg_error
 from tawazi.config import cfg
 from tawazi.consts import (
     ARG_NAME_ACTIVATE,
@@ -32,6 +32,7 @@ from tawazi.consts import (
     XNOutsideDAGCall,
 )
 from tawazi.errors import TawaziBaseException, TawaziUsageError
+from tawazi.node.uxn import UsageExecNode
 from tawazi.profile import Profile
 
 from .helpers import _lazy_xn_id, _validate_tuple, make_suffix
@@ -467,288 +468,3 @@ class LazyExecNode(ExecNode, Generic[P, RVXN]):
             # https://stackoverflow.com/questions/3798835/understanding-get-and-set-and-python-descriptors
             return self
         return MethodType(self, instance)  # func=self  # obj=instance
-
-
-def _xn(func: Callable[P, RVXN]) -> LazyExecNode[P, RVXN]:
-    return LazyExecNode(
-        exec_function=func,
-        priority=0,
-        is_sequential=cfg.TAWAZI_IS_SEQUENTIAL,
-        debug=False,
-        tag=None,
-        setup=False,
-        unpack_to=None,
-        resource=cfg.TAWAZI_DEFAULT_RESOURCE,
-    )
-
-
-# boolean operators definitions
-# "rich comparison" methods definitions
-@_xn
-def _lt(a: Any, b: Any) -> bool:
-    return a < b  # type: ignore[no-any-return]
-
-
-@_xn
-def _le(a: Any, b: Any) -> bool:
-    return a <= b  # type: ignore[no-any-return]
-
-
-@_xn
-def _eq(a: Any, b: Any) -> bool:
-    return a == b  # type: ignore[no-any-return]
-
-
-@_xn
-def _ne(a: Any, b: Any) -> bool:
-    return a != b  # type: ignore[no-any-return]
-
-
-@_xn
-def _gt(a: Any, b: Any) -> bool:
-    return a > b  # type: ignore[no-any-return]
-
-
-@_xn
-def _ge(a: Any, b: Any) -> bool:
-    return a >= b  # type: ignore[no-any-return]
-
-
-# "numeric emulation" methods definitions
-# https://docs.python.org/3/reference/datamodel.html#emulating-numeric-types
-
-# binary operations
-# automatically implement:
-# 1. right hand operations (allowing float + UsageExecNode[float])
-# 2. augmented arithmetic (+=, -=, etc.)@_xn
-
-
-@_xn
-def _add(a: Any, b: Any) -> Any:
-    return a + b
-
-
-@_xn
-def _sub(a: Any, b: Any) -> Any:
-    return a - b
-
-
-@_xn
-def _mul(a: Any, b: Any) -> Any:
-    return a * b
-
-
-@_xn
-def _matmul(a: Any, b: Any) -> Any:
-    return a @ b
-
-
-@_xn
-def _truediv(a: Any, b: Any) -> Any:
-    return a / b
-
-
-@_xn
-def _floordiv(a: Any, b: Any) -> Any:
-    return a // b
-
-
-@_xn
-def _mod(a: Any, b: Any) -> Any:
-    return a % b
-
-
-@_xn
-def _divmod(a: Any, b: Any) -> Any:
-    return divmod(a, b)
-
-
-@_xn
-def _pow(a: Any, b: Any) -> Any:
-    return pow(a, b)
-
-
-@_xn
-def _lshift(a: Any, b: Any) -> Any:
-    return a << b
-
-
-@_xn
-def _rshift(a: Any, b: Any) -> Any:
-    return a >> b
-
-
-@_xn
-def _and(a: Any, b: Any) -> Any:
-    return a & b
-
-
-@_xn
-def _xor(a: Any, b: Any) -> Any:
-    return a ^ b
-
-
-@_xn
-def _or(a: Any, b: Any) -> Any:
-    return a | b
-
-
-# unary operations
-@_xn
-def _neg(a: Any) -> Any:
-    return -a
-
-
-@_xn
-def _pos(a: Any) -> Any:
-    return +a
-
-
-@_xn
-def _abs(a: Any) -> Any:
-    return abs(a)
-
-
-@_xn
-def _invert(a: Any) -> Any:
-    return ~a
-
-
-# built ins
-# when a value is used for indexing of another
-# can't be implemented because should return appropriate type
-# object.__index__(self)
-
-# can't be implemented because should return appropriate type
-# __complex__
-# __int__
-# __float__
-
-# __round__
-# __trunc__
-# __floor__
-# __ceil__
-
-
-# NOTE: None is hashable! In theory it can be used as a key in a dict!
-KeyType = Union[str, int, Tuple[Any, ...], None, NoValType]
-
-
-# TODO: transform this logic into the ExecNode itself ?
-@dataclass(frozen=True)
-class UsageExecNode:
-    """The usage of the ExecNode / LazyExecNode inside the function describing the DAG.
-
-    If ExecNode is not indexed with a key or an int, NoVal is used as the key.
-    """
-
-    id: Identifier
-    key: List[KeyType] = field(default_factory=list)
-
-    # TODO: make type of key immutable or something hashable, that way
-    #  we can directly build a nx DAG with nodes instead of node ids
-    #  removing almost all of the duplicate graph building logic
-    # used in the dag dependency description
-    def __getitem__(self, key: KeyType) -> "UsageExecNode":
-        """Record the used key in a new UsageExecNode.
-
-        Args:
-            key (Union[str, int, Tuple[Any]]): the used key for indexing (whether int like Lists or strings like dicts)
-
-        Returns:
-            UsageExecNode: the new UsageExecNode where the key is recorded
-        """
-        # deepcopy self because UsageExecNode can be reused with different indexing
-        new = deepcopy(self)
-        new.key.append(key)
-        return new
-
-    @property
-    def is_indexable(self) -> bool:
-        """Whether UsageExecNode is used with an index.
-
-        Returns:
-            bool: whether the ExecNode is indexable
-        """
-        return bool(self.key)
-
-    def result(self, xn_dict: Dict[Identifier, ExecNode]) -> Any:
-        """Extract the result of the ExecNode corresponding to used key(s).
-
-        Returns:
-            Any: value inside the container
-        """
-        xn = xn_dict[self.id]
-        # ignore typing error because it is the responsibility of the user to ensure the result of the XN is indexable!
-        # Will raise the appropriate exception automatically
-        #  The user might have specified a subgraph to run => xn contain NoVal
-        #  or the user tried to access a non-indexable object
-        # NOTE: maybe handle the 3 types of exceptions that might occur properly to help the user through debugging
-        # if isinstance(xn.result, NoValType):
-        #     raise TawaziTypeError(f"{xn} didn't run, hence its result is not indexable. Check your DAG's config")
-
-        return _filter_noval(reduce(lambda obj, key: obj.__getitem__(key), self.key, xn.result))
-
-    def __bool__(self) -> NoReturn:
-        """__bool__ operator."""
-        # __bool__ can not be faked because it must return True or False
-        raise NotImplementedError
-
-    def __contains__(self, other: Any) -> NoReturn:
-        """___contains__ operator."""
-        # __contains__ can not be faked because the return value gets automatically casted to bool
-        raise NotImplementedError
-
-
-def reflected(operator: Callable[[Any, Any], Any]) -> Callable[[Any, Any], Any]:
-    """Reflects an operator."""
-
-    def inner_reflected(a: Any, b: Any) -> Any:
-        return operator(b, a)
-
-    return inner_reflected
-
-
-# binary operations
-setattr(UsageExecNode, "__lt__", _lt)  # noqa: B010
-setattr(UsageExecNode, "__le__", _le)  # noqa: B010
-setattr(UsageExecNode, "__eq__", _eq)  # noqa: B010
-setattr(UsageExecNode, "__ne__", _ne)  # noqa: B010
-setattr(UsageExecNode, "__gt__", _gt)  # noqa: B010
-setattr(UsageExecNode, "__ge__", _ge)  # noqa: B010
-setattr(UsageExecNode, "__add__", _add)  # noqa: B010
-setattr(UsageExecNode, "__sub__", _sub)  # noqa: B010
-setattr(UsageExecNode, "__mul__", _mul)  # noqa: B010
-setattr(UsageExecNode, "__matmul__", _matmul)  # noqa: B010
-setattr(UsageExecNode, "__truediv__", _truediv)  # noqa: B010
-setattr(UsageExecNode, "__floordiv__", _floordiv)  # noqa: B010
-setattr(UsageExecNode, "__mod__", _mod)  # noqa: B010
-setattr(UsageExecNode, "__divmod__", _divmod)  # noqa: B010
-setattr(UsageExecNode, "__pow__", _pow)  # noqa: B010
-setattr(UsageExecNode, "__lshift__", _lshift)  # noqa: B010
-setattr(UsageExecNode, "__rshift__", _rshift)  # noqa: B010
-setattr(UsageExecNode, "__and__", _and)  # noqa: B010
-setattr(UsageExecNode, "__xor__", _xor)  # noqa: B010
-setattr(UsageExecNode, "__or__", _or)  # noqa: B010
-
-# reflected binary operators
-setattr(UsageExecNode, "__radd__", reflected(_add))  # noqa:B010
-setattr(UsageExecNode, "__rsub__", reflected(_sub))  # noqa:B010
-setattr(UsageExecNode, "__rmul__", reflected(_mul))  # noqa:B010
-setattr(UsageExecNode, "__rmatmul__", reflected(_matmul))  # noqa:B010
-setattr(UsageExecNode, "__rtruediv__", reflected(_truediv))  # noqa:B010
-setattr(UsageExecNode, "__rfloordiv__", reflected(_floordiv))  # noqa:B010
-setattr(UsageExecNode, "__rmod__", reflected(_mod))  # noqa:B010
-setattr(UsageExecNode, "__rdivmod__", reflected(_divmod))  # noqa:B010
-setattr(UsageExecNode, "__rpow__", reflected(_pow))  # noqa:B010
-setattr(UsageExecNode, "__rlshift__", reflected(_lshift))  # noqa:B010
-setattr(UsageExecNode, "__rrshift__", reflected(_rshift))  # noqa:B010
-setattr(UsageExecNode, "__rand__", reflected(_and))  # noqa:B010
-setattr(UsageExecNode, "__rxor__", reflected(_xor))  # noqa:B010
-setattr(UsageExecNode, "__ror__", reflected(_or))  # noqa:B010
-
-# unary operations
-setattr(UsageExecNode, "__neg__", _neg)  # noqa: B010
-setattr(UsageExecNode, "__pos__", _pos)  # noqa: B010
-setattr(UsageExecNode, "__abs__", _abs)  # noqa: B010
-setattr(UsageExecNode, "__invert__", _invert)  # noqa: B010
